# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: NOMOS API
  version: 0.1.0
  description: |
    에이전트 협업 조율 서버. 수동 테스트 절차는 `docs/manual-test.md`를 함께 보라.

    **토큰이 두 종류다.** 오른쪽 위 Authorize를 누르면 칸이 두 개 뜬다.

    | 칸 | 어떤 토큰 | 받는 법 |
    |---|---|---|
    | `userToken` | 사람 토큰 (`kind: user`) | `POST /auth/login` — 시드 계정 `rep` / `manual-test-1234` |
    | `agentToken` | 에이전트 토큰 (`kind: agent`) | `npm run seed` 출력의 **be-laptop access token**, 또는 `POST /agents/token/refresh` |

    서로의 API를 부를 수 없다. 사람 토큰으로 태스크 API를 부르면 401이다.

    access token은 **1시간**이다. 만료되면 `POST /agents/token/refresh`로 다시 받는다.
servers:
  - url: http://localhost:3000/api
    description: 로컬 개발 서버
components:
  securitySchemes:
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 사람 토큰. `POST /auth/login`의 `data.accessToken`을 붙여넣는다.
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 에이전트 토큰. 시드 출력의 access token 또는 refresh 응답의 값을 붙여넣는다.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: FORBIDDEN_PATH
            message:
              type: string
              example: '**/.env* is denied'
            reason:
              type: string
              description: POLICY_STALE에만 실린다. 다른 코드에서는 생략된다.
              example: policy_stale
  responses:
    Unauthenticated:
      description: 토큰이 없거나 만료·위조됨
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: { code: UNAUTHENTICATED, message: invalid or expired token }
    PolicyStale:
      description: |
        토큰이 **옛 정책 스냅샷 기준**이다. 경로 규칙이나 정책이 바뀌면 이렇게 된다.
        refresh로 재발급하고 원 요청을 **1회만** 재시도한다. 그래도 stale이면 멈춘다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: POLICY_STALE
              message: policy snapshot changed; refresh the token
              reason: policy_stale
...
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{TASK_BE}}'
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{ORG_ID}}'
...
paths:
  /auth/signup:
    post:
      tags: [인증·계정]
      summary: 로컬 계정 가입
      description: |
        연결 키 평문은 **이 응답에서 한 번만** 나온다. 이후 다시 볼 방법이 없고, 잃어버리면 재발급해야 한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginId, password, nickname]
              properties:
                loginId: { type: string, minLength: 3, maxLength: 64 }
                password: { type: string, minLength: 8 }
                nickname: { type: string }
            example:
              loginId: tester
              password: manual-test-1234
              nickname: tester
      responses:
        '201':
          description: 가입됨
          content:
            application/json:
              example:
                data:
                  userId: 3f2a1b4c-0000-0000-0000-000000000000
                  connectKey: lt4mWBZ-1_bw93HtjUQdKr6_U7TLjI8dpmsRI8Fu-D8
        '409':
          description: 아이디 중복
          content:
            application/json:
              example:
                error: { code: LOGIN_ID_TAKEN, message: login id is already taken }
  /auth/login:
    post:
      tags: [인증·계정]
      summary: 로그인 (사람 토큰 발급)
      description: |
        **없는 아이디와 틀린 비밀번호는 같은 응답이다.** 없는 아이디일 때도 scrypt를 한 번 돌려
        응답 시간까지 맞춘다 — 아이디 존재 여부가 드러나지 않게.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginId, password]
              properties:
                loginId: { type: string }
                password: { type: string }
            example:
              loginId: rep
              password: manual-test-1234
      responses:
        '200':
          description: 토큰 발급. `data.accessToken`을 Authorize의 userToken 칸에 붙여넣는다.
          content:
            application/json:
              example:
                data: { accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..., tokenType: Bearer, expiresIn: 3600 }
        '401':
          description: 아이디 또는 비밀번호가 틀림
          content:
            application/json:
              example:
                error: { code: INVALID_CREDENTIALS, message: invalid login id or password }
  /me/connect-key/rotate:
    post:
      tags: [인증·계정]
      summary: 개인 연결 키 재발급
      description: 새 키를 발급하면 **기존 키는 즉시 무효**다.
      security: [{ userToken: [] }]
      responses:
        '200':
          description: 새 연결 키 (평문은 이 응답에서 한 번만)
          content:
            application/json:
              example:
                data: { connectKey: o9m9J-utE-VIKTelv9j8AAAiVTDPHX0d7ikdFu48aig }
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /agents/connect:
    post:
      tags: [CLI 연결]
      summary: 에이전트 등록 (연결 키로)
      description: |
        조직·프로젝트가 없어도 연결된다(`status=pending`). 같은 `agentName`으로 다시 부르면
        새로 만들지 않고 갱신한다 — CLI 재설치 대응.

        **틀린 연결 키는 401이 아니라 400 + 일반 메시지다.** 무엇이 틀렸는지 알려주지 않는다.

        연결 키는 서버에 **해시로만** 저장되므로 아래 example을 DB에서 채워줄 수 없다.
        `npm run seed` 출력의 값을 직접 붙여넣어야 한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connectKey, agentName, harness]
              properties:
                connectKey: { type: string }
                agentName: { type: string, maxLength: 64 }
                harness: { type: string }
                skills: { type: array, items: { type: string } }
                maxConcurrent: { type: integer, minimum: 1, maximum: 16 }
            example:
              connectKey: 시드-출력의-be-연결-키를-붙여넣으세요
              agentName: be-laptop
              harness: claude-code@2.1.263
              skills: [typescript]
              maxConcurrent: 2
      responses:
        '201':
          description: 등록됨. access는 1시간, refresh는 30일.
          content:
            application/json:
              example:
                data: { agentId: 2012e1df-0beb-43fd-ac09-d893ba56b01e, accessToken: eyJ..., refreshToken: XtvVbBmp... }
        '400':
          description: 연결 키가 틀림 (의도적으로 401이 아니다)
          content:
            application/json:
              example:
                error: { code: INVALID_CONNECT_REQUEST, message: invalid connect request }
  /agents/token/refresh:
    post:
      tags: [CLI 연결]
      summary: 에이전트 access 토큰 재발급
      description: |
        **401 POLICY_STALE을 받으면 여기로 재발급하고 원 요청을 1회만 재시도한다.**
        재발급 후에도 stale이면 루프이므로 즉시 멈춘다.

        재발급된 토큰에는 현재 `project_id`와 `policy_hash`가 담긴다 — 프로젝트에 배정된 뒤
        이걸 한 번 불러야 태스크 API를 쓸 수 있다.

        refresh 토큰도 해시로만 저장되므로 example을 DB에서 채울 수 없다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
            example:
              refreshToken: 시드-출력의-be-laptop-refresh-token을-붙여넣으세요
      responses:
        '200':
          description: 새 access 토큰
          content:
            application/json:
              example:
                data: { accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
        '401':
          description: refresh 토큰이 만료·폐기됨
          content:
            application/json:
              example:
                error: { code: INVALID_REFRESH_TOKEN, message: refresh token is invalid or expired }
...
  /orgs/{orgId}/invites:
    post:
      tags: [초대]
      summary: 초대 링크 발급 (대표 전용)
      description: 토큰은 이벤트 payload에 남기지 않는다 — events는 지워지지 않으므로 비밀값을 넣지 않는다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teamRole: { type: string, enum: [FRONTEND, BACKEND] }
                expiresInDays: { type: integer, minimum: 1 }
            example:
              teamRole: BACKEND
              expiresInDays: 7
      responses:
        '201':
          description: 발급됨
          content:
            application/json:
              example:
                data: { token: '{{INVITE_TOKEN}}', url: 'http://localhost:5173/invites/{{INVITE_TOKEN}}', expiresAt: '2026-09-25T00:00:00.000Z', teamRole: BACKEND }
        '403':
          description: 대표가 아님
          content:
            application/json:
              example:
                error: { code: NOT_REPRESENTATIVE, message: this action requires the organization representative }
  /invites/{token}:
    get:
      tags: [초대]
      summary: 수락 전 미리보기 (인증 불필요)
      description: |
        **무효한 토큰도 404가 아니라 200 + `valid:false`다.** 404를 쓰면 토큰 존재 여부가 드러난다.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
          example: '{{INVITE_TOKEN}}'
      responses:
        '200':
          description: 유효하든 아니든 200
          content:
            application/json:
              examples:
                유효:
                  value: { data: { valid: true, orgName: Acme Inc., teamRole: BACKEND, expiresAt: '2026-09-25T00:00:00.000Z' } }
                무효:
                  value: { data: { valid: false } }
  /invites/{token}/accept:
    post:
      tags: [초대]
      summary: 초대 수락
      description: |
        `users.org_id`와 그 사용자의 `agents.org_id`가 **함께** 채워진다.
        같은 링크를 다시 눌러도 성공한다(멱등). 이미 다른 조직에 속했으면 409.
      security: [{ userToken: [] }]
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
          example: '{{INVITE_TOKEN}}'
      responses:
        '200':
          description: 합류됨
          content:
            application/json:
              example:
                data: { userId: 2a5f7d0d-60ce-4937-a67a-a567a41c8ddb, orgId: '{{ORG_ID}}' }
        '409':
          description: 이미 다른 조직 소속
          content:
            application/json:
              example:
                error: { code: ALREADY_IN_ORG, message: user already belongs to an organization }
        '410':
          description: 만료되거나 이미 사용됨
          content:
            application/json:
              example:
                error: { code: INVITE_EXPIRED, message: invite has expired }
...
  /tasks/{taskId}/notes:
    post:
      tags: [태스크·노트]
      summary: publish_note — 인계 노트 발행
      description: |
        노트는 **대화 채널이 아니라 기록**이다. 본문 컬럼도, 수신자 지정도, 답글도 없다.
        정정은 새 행(`supersedes`)이다.

        **제목은 서버가 조립해** 응답의 `title`에 담는다: `#순번 - 역할 기능키 종류 — headline`.

        형식 위반은 **자르지 않고 422**로 어디가 몇 자인지 돌려준다. `DEVIATION`은 `affects`가 필수다.
        다른 에이전트의 이름·id를 넣으면 거부하고 `raise_dispute`로 안내한다 —
        **자연어로 심사하지 않는다.** 이름 없이 비난하는 문장은 통과한다.
      security: [{ agentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, headline, keyPoints]
              properties:
                kind: { type: string, enum: [IMPLEMENTED, DECIDED, GOTCHA, DEVIATION] }
                headline: { type: string, maxLength: 60 }
                keyPoints: { type: array, minItems: 1, maxItems: 5, items: { type: string, maxLength: 120 } }
                affects: { type: array, maxItems: 10, items: { type: string, maxLength: 200 } }
                supersedes: { type: string, format: uuid }
            examples:
              구현완료:
                value: { kind: IMPLEMENTED, headline: participation API done, keyPoints: ['409 on full study'] }
              명세이탈:
                value: { kind: DEVIATION, headline: quota check moved to service, keyPoints: [handler was too late], affects: [contracts/F-03.yaml] }
              위반예시_affects누락:
                value: { kind: DEVIATION, headline: x, keyPoints: [y] }
      responses:
        '201':
          description: 발행됨
          content:
            application/json:
              example:
                data:
                  id: 55555555-5555-5555-5555-555555555555
                  seq: 1
                  kind: IMPLEMENTED
                  headline: participation API done
                  keyPoints: ['409 on full study']
                  affects: []
                  title: '#1 - 백엔드 F-03 구현 완료 — participation API done'
        '401':
          $ref: '#/components/responses/PolicyStale'
        '403':
          description: CLAIM하지 않은 태스크
          content:
            application/json:
              example:
                error: { code: NOT_TASK_ASSIGNEE, message: task is assigned to another agent }
        '422':
          description: |
            형식 위반. **권한 거부가 아니므로 `TOOL_DENIED`를 남기지 않는다** —
            오타를 차단 지표(M5′)에 섞지 않기 위해서다.

            위반 목록은 자르지 않고 `error.details`에 전부 담는다. 어느 필드의 몇 번째가
            몇 자인지 모르면 에이전트가 같은 요청을 반복한다.
          content:
            application/json:
              example:
                error:
                  code: NOTE_INVALID
                  message: note validation failed
                  details:
                    - { field: keyPoints, index: 0, length: 121, limit: 120, message: key_points의 각 항목은 120자 이하여야 합니다 }
                    - { field: kind, message: DEVIATION은 affects가 1개 이상이어야 합니다 }
  /projects/{projectId}/notes:
    get:
      tags: [태스크·노트]
      summary: read_notes — 인계 노트 읽기
      description: |
        **이벤트를 남기지 않는다.** 조회까지 기록하면 events가 열람 로그가 되고,
        리플레이에서 "무엇이 상태를 바꿨나"를 찾기 어려워진다.

        내 태스크에 필요한 노트는 서버가 프롬프트에 주입하므로, 에이전트는 더 필요할 때만 쓴다.

        **사람도 호출한다** — 노트는 "무엇을 어떻게 진행했는가"의 기록이므로 웹 UI의 진행 상황
        화면의 본문이 된다. 사람은 대표이거나 그 프로젝트에 배정된 에이전트의 주인이어야 한다.
        단, 노트는 에이전트의 **자기 보고**다. 검증 통과 여부는 `verifications`에서 확인한다.
      security: [{ agentToken: [] }, { userToken: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
          example: '{{PROJECT_ID}}'
        - name: spec_id
          in: query
          schema: { type: string, format: uuid }
          description: 이 기능 명세의 노트만
        - name: since_seq
          in: query
          schema: { type: integer, minimum: 0 }
          description: 이 순번 이후만
          example: 0
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
          example: 5
      responses:
        '200':
          description: 최신순 목록
          content:
            application/json:
              example:
                data:
                  notes:
                    - { id: 55555555-5555-5555-5555-555555555555, seq: 1, kind: IMPLEMENTED, headline: participation API done, affects: [] }
        '401':
          $ref: '#/components/responses/PolicyStale'
        '403':
          description: 다른 프로젝트
          content:
            application/json:
              example:
                error: { code: NOT_PROJECT_MEMBER, message: agent is not a member of this project }
...
  /agents/me:
    get:
      tags: [CLI 연결]
      summary: 에이전트 자기 설정
      description: |
        Executor가 기동할 때 자기 프로젝트·역할·동시 실행 상한을 읽는다. 이 값들을 Executor 쪽 설정
        파일에 적어두면 서버에서 바꿔도 반영되지 않으므로, 서버가 정본을 준다.
      security: [{ agentToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                data:
                  agentId: 2012e1df-0beb-43fd-ac09-d893ba56b01e
                  projectId: '{{PROJECT_ID}}'
                  teamRole: BACKEND
                  maxConcurrent: 2
...